+++ Service/Clustering/pom.xml
@@ -12,5 +12,25 @@
     <groupId>com.hzgc.service</groupId>
     <artifactId>clustering</artifactId>
 
+    <dependencies>
+        <dependency>
+            <groupId>com.hzgc.service</groupId>
+            <artifactId>bigdata-api</artifactId>
+            <version>${project.parent.version}</version>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>com.hzgc.service</groupId>
+            <artifactId>util</artifactId>
+            <version>${project.parent.version}</version>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>com.hzgc.service</groupId>
+            <artifactId>dynrepo</artifactId>
+            <version>${project.parent.version}</version>
+            <scope>provided</scope>
+        </dependency>
+    </dependencies>
 
 </project>
