--- Service/DynRepo/src/mian/java/com/hzgc/dynamicrepo/CaptureStatuCode.java
@@ -1,14 +1,0 @@
-package com.hzgc.dynamicrepo;
-
-public enum  CaptureStatuCode {
-    THRESHOLDNULL("FEATURENULL");
-    private String code;
-    CaptureStatuCode(String code) {
-        this.code = code;
-    }
-
-    @Override
-    public String toString() {
-        return code;
-    }
-}
