+++ Service/Util/src/main/java/com/hzgc/service/util/Constants.java
@@ -1,0 +1,7 @@
+package com.hzgc.service.util;
+
+/**
+ * service 模块常量定义
+ */
+public class Constants {
+}
