+++ Service/Util/src/main/java/com/hzgc/service/util/UuidUtil.java
@@ -1,0 +1,20 @@
+package com.hzgc.service.util;
+
+import java.io.Serializable;
+import java.util.UUID;
+
+public class UuidUtil implements Serializable {
+    private String uuid;
+
+    public static String setUuid() {
+        return UUID.randomUUID().toString().replaceAll("-", "");
+    }
+
+    public String getUuid() {
+        return uuid;
+    }
+
+    public void setUuid(String uuid) {
+        this.uuid = uuid;
+    }
+}
