+++ Service/Util/src/main/resources/es-config.properties
@@ -1,0 +1,4 @@
+es.cluster.name=hbase2es-cluster
+## only need one of the es nodes of es cluster
+es.hosts=172.18.18.103
+es.cluster.port=9300
