+++ Service/Util/src/main/resources/hbase-site.xml
@@ -1,0 +1,91 @@
+<?xml version="1.0"?>
+<?xml-stylesheet type="text/xsl" href="configuration.xsl"?>
+<!--
+/**
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+-->
+<configuration>
+    <property>
+        <name>hbase.rootdir</name>
+	<value>hdfs://hzgc/hbase</value>
+    </property>
+    <property>
+        <name>hbase.master.port</name>
+        <value>60000</value>
+    </property>
+    <property>
+        <name>hbase.regionserver.port</name>
+        <value>16020</value>
+    </property>
+    <property>
+        <name>hbase.regionserver.info.port</name>
+        <value>16030</value>
+    </property>
+    <property>
+        <name>hbase.master.info.port</name>
+        <value>60010</value>
+    </property>
+    <property>
+        <name>hbase.cluster.distributed</name>
+        <value>true</value>
+    </property>
+    <property>
+        <name>hbase.zookeeper.property.clientPort</name>
+        <value>2181</value>
+    </property>
+    <property>
+        <name>hbase.zookeeper.quorum</name>
+        <value>s108:2181,s104:2181,s103:2181</value>
+    </property>
+    <property>
+        <name>hbase.tmp.dir</name>
+        <value>/opt/hzgc/bigdata/HBase/hbase/tmp</value>
+    </property>
+    <property>
+        <name>hbase.zookeeper.property.dataDir</name>
+        <value>/opt/hzgc/bigdata/HBase/hbase/hbase_zk_datadir</value>
+    </property>
+    <property>
+        <name>hbase.master.maxclockskew</name>
+        <value>300000</value>
+    </property>
+    <property>
+        <name>dfs.replication</name>
+        <value>3</value>
+    </property>
+    <!-- 新增的Phoenix配置 -->
+    <property>
+        <name>phoenix.functions.allowUserDefinedFunctions</name>
+        <value>true</value>
+    </property>
+    <property>
+        <name>fs.hdfs.impl</name>
+        <value>org.apache.hadoop.hdfs.DistributedFileSystem</value>
+    </property>
+    <property>
+        <name>hbase.dynamic.jars.dir</name>
+        <value>${hbase.rootdir}/lib</value>
+        <description>
+            The directory from which the custom udf jars can be loaded
+            dynamically by the phoenix client/region server without the need to restart. However,
+            an already loaded udf class would not be un-loaded. See
+            HBASE-1936 for more details.
+        </description>
+    </property>
+</configuration>
