+++ Service/Util/src/main/resources/jdbc.properties
@@ -1,0 +1,23 @@
+#连接设置
+driverClassName=org.apache.hive.jdbc.HiveDriver
+url=jdbc:hive2://s103:2181,s104:2181,s108:2181/;serviceDiscoveryMode=zooKeeper;zooKeeperNamespace=thriftserver
+#<!-- 初始化连接 -->
+initialSize=10
+#最大连接数量
+maxActive=50
+#<!-- 最小空闲连接 -->
+minIdle=5
+#<!-- 超时等待时间以毫秒为单位 60000毫秒/1000等于60秒 -->
+maxWait=60000
+#JDBC驱动建立连接时附带的连接属性属性的格式必须为这样：[属性名=property;]
+#注意："user" 与 "password" 两个属性会被明确地传递，因此这里不需要包含他们。
+#connectionProperties=useUnicode=true;characterEncoding=utf-8
+#指定由连接池所创建的连接的自动提交（auto-commit）状态。
+defaultAutoCommit=true
+#driver default 指定由连接池所创建的连接的只读（read-only）状态。
+#如果没有设置该值，则“setReadOnly”方法将不被调用。（某些驱动并不支持只读模式，如：Informix）
+defaultReadOnly=
+#driver default 指定由连接池所创建的连接的事务级别（TransactionIsolation）。
+testWhileIdle=false
+
+phoenixJDBCURL=172.18.18.103:2181
