+++ Service/pom.xml
@@ -28,6 +28,12 @@
             <artifactId>elasticsearch</artifactId>
             <version>1.0</version>
         </dependency>
+        <dependency>
+            <groupId>com.hzgc.cluster</groupId>
+            <artifactId>hbase</artifactId>
+            <version>${project.parent.version}</version>
+            <scope>provided</scope>
+        </dependency>
     </dependencies>
 
 </project>
